module6/pyflink/requirements.txt
apache-flink==1.16.0
psycopg2-binary==2.9.1
requests
kafka-python
protobuf==3.17.3
